+++ InternProj/Models/XuatNhapTonData.cs
@@ -10,13 +10,13 @@ namespace InternProj.Models
 
         public string TenSP { get; set; } = string.Empty;
 
-        public int SoLuongDauKy { get; set; }
+        public decimal SoLuongDauKy { get; set; }
 
-        public int SoLuongNhap { get; set; }
+        public decimal SoLuongNhap { get; set; }
 
-        public int SoLuongXuat { get; set; }
+        public decimal SoLuongXuat { get; set; }
 
-        public int SoLuongCuoiKy { get; set; }
+        public decimal SoLuongCuoiKy { get; set; }
         public override string ToString() => $"{MaSP}";
     }
 }
